Step 7: move JobLogs into the meeseeks package
